backend/requirements.txt
alembic==1.18.4
fastapi==0.135.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
sqlmodel==0.0.37
uvicorn==0.42.0
pytest==8.4.2
httpx==0.28.1
PyJWT==2.10.1
pwdlib[argon2]==0.3.0